+++ src/common/Definitions.h
@@ -30,45 +30,47 @@ Default font for CinnamonToast to use.
 /**
  * Internal app version for compatibility checks.
  */
-#define APP_INTERNAL_VERSION 0xC000007L
+#define APP_INTERNAL_VERSION 0x00000008L
 
 /**
  * External app version to be shown to the end user.
  */
-#define APP_VERSION "1.5.0+snapshot-2025c01a"
+#define APP_VERSION "1.5.0"
 
 /**
  * Error code specification
  *
- * These error codes should be used to indicate a specific error. Some errors
- * are platform-specific, and this is indicated by the second leftmost digit of
- * the error code. The valid values are: 0x0 - Cross-platform 0x1 - Windows 0x2
+ * These ctoastError codes should be used to indicate a specific ctoastError.
+ * Some ctoastErrors are platform-specific, and this is indicated by the second
+ * leftmost digit of the ctoastError code. The valid values are: 0x0 -
+ * Cross-platform 0x1 - Windows 0x2
  * - macOS 0x3 - Linux
  *
- * Examples are: 0x01000000 (an "error" code for a successful execution on
+ * Examples are: 0x01000000 (an "ctoastError" code for a successful execution on
  * Windows, though this is only for demonstation purposes and SHOULD NOT be used
  * in code), 0x03badf00 (Linux), etc.
  *
  *
  * There are some reserved return codes like 0x0000002 which is ALWAYS the "file
- * not found" error. 0x00000000 is also another one of these codes. It's not
- * necessarily an "error" code, though it is a return code for success.
+ * not found" ctoastError. 0x00000000 is also another one of these codes. It's
+ * not necessarily an "ctoastError" code, though it is a return code for
+ * success.
  */
 
 /**
- * Comment/uncomment this to toggle Win32 errors.
+ * Comment/uncomment this to toggle Win32 ctoastErrors.
  */
 #ifdef _WIN32
-// Define Win32 error codes
+// Define Win32 ctoastError codes
 #define WIN32_ERROR_CODES
 #endif
 
-// Internal error codes and other things
+// Internal ctoastError codes and other things
 
 // Error for if there isn't a file specified.
 #define CTOAST_ERROR_NO_FILES_SPECIFIED 0x000003f0L
 
-// Error for an XML error that does not originate from the program.
+// Error for an XML ctoastError that does not originate from the program.
 #define CTOAST_ERROR_GENERIC_XML_ERROR 0x000003f1L
 
 // Error for if the XML file does not have a root element.
@@ -87,11 +89,11 @@ Default font for CinnamonToast to use.
 // Error if a hex color (like #ffffff) is malformed.
 #define CTOAST_ERROR_HEX_COLOR_MALFORMED 0x000003f4L
 
-// A Linux-specific error that X11 isn't initialized.
+// A Linux-specific ctoastError that X11 isn't initialized.
 #define CTOAST_ERROR_LINUX_X11_NOT_INITIALIZED 0x030003f5L
 
 // Error if the shared library (.dll, .so, etc.) can't be loaded primarily due
-// to filesystem errors or other reasons.
+// to filesystem ctoastErrors or other reasons.
 #define CTOAST_ERROR_CANNOT_LOAD_SHARED_LIBRARY 0x000003f9L
 
 // Error if the shared library's `CToastMain` function does not exist or there
@@ -104,11 +106,11 @@ Default font for CinnamonToast to use.
 // Error if an unhandled exception (std::exception) occurs.
 #define CTOAST_ERROR_UNHANDLED_EXCEPTION 0x000003fcL
 
-// Win32 specific errors
+// Win32 specific ctoastErrors
 #if defined(_WIN32) && defined(WIN32_ERROR_CODES)
 
 // Error if the parent HWND (window handle) is invalid in Windows.
 #define CTOAST_ERROR_WIN_PARENT_HWND_INVALID 0x010003f5L
 
-// more errors go here
+// more ctoastErrors go here
 #endif