+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 # v1.2.0 (2025-03-24)
 - Added support for Lua scripting
 - Fixed some bugs
-- Added some error codes
+- Added some ctoastError codes
 - Added a whole new Lua API for interfacing with the C++ code
 - Added a new module (lua) for handling Lua instances
 
@@ -23,7 +23,7 @@
 - Added shared library APIs (rudimentary)
 - Added support for DLL loading (for code injection)
 - Revamped .xml format to support shared libraries
-- Added some error codes
+- Added some ctoastError codes
 
 # v1.0.0 (unknown date)
 - Initial release